refactor(today): TodayView 拆分第一批 — 抽出 598 行独立子视图/modifier
#16 密度收敛。把 TodayView.swift 末尾自成一体、与 TodayView @State 零耦合的
辅助类型整体搬到两个语义清晰的新文件（纯重构，零行为变化）：

- TodayViewComponents.swift（521 行）— DayNavTarget / TimelineShareText /
  CompilationFailedBanner / LocationDraftCard(+Row) / CompilationProgressBar /
  TimelineRow，都是通过 let/闭包参数化的展示子视图与数据包装
- TodayViewModifiers.swift（103 行）— ScrollOffsetPreferenceKey /
  ScrollEntranceModifier / TodayNumericTextTransition / CardZoomSource /
  CardZoomDestination，纯 ViewModifier + PreferenceKey

跨文件后原 file-private 的 ScrollOffsetPreferenceKey / NumericTextContentTransition
需放宽到 internal；后者与 Search/Graph 里各自的 private 同名副本会撞名，故本文件
这份改名 TodayNumericTextTransition 以避免波及无关文件（DRY 统一留待专门 PR）。

TodayView.swift 4084 → 3487 行。BUILD SUCCEEDED（iPhone 17）。
